rename constants files
