+++ berkeley-function-call-leaderboard/bfcl/_apply_function_credential_config.py
@@ -2,7 +2,7 @@ import glob
 import json
 import os
 
-from bfcl.constant import PROMPT_PATH
+from bfcl.constants import PROMPT_PATH
 from bfcl.eval_checker.executable_eval.custom_exception import NoAPIKeyError
 
 ENV_VARS = ("GEOCODE_API_KEY", "RAPID_API_KEY", "OMDB_API_KEY", "EXCHANGERATE_API_KEY")
