+++ berkeley-function-call-leaderboard/bfcl/eval_checker/ast_eval/type_convertor/js_type_converter.py
@@ -1,5 +1,5 @@
 import re
-from bfcl.model_handler.constant import JS_TYPE_CONVERSION
+from bfcl.model_handler.function_calling_constants import JS_TYPE_CONVERSION
 
 
 def js_type_converter(value, expected_type, nested_type=None):
