+++ berkeley-function-call-leaderboard/bfcl/model_handler/base_handler.py
@@ -2,13 +2,13 @@ import json
 import time
 from copy import deepcopy
 
-from bfcl.constant import RESULT_PATH, VERSION_PREFIX
+from bfcl.constants import RESULT_PATH, VERSION_PREFIX
 from bfcl.eval_checker.multi_turn_eval.multi_turn_utils import (
     STATELESS_CLASSES,
     execute_multi_turn_func_call,
     is_empty_execute_response,
 )
-from bfcl.model_handler.constant import (
+from bfcl.model_handler.function_calling_constants import (
     DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_FC,
     DEFAULT_USER_PROMPT_FOR_ADDITIONAL_FUNCTION_PROMPTING,
     MAXIMUM_STEP_LIMIT,
