+++ berkeley-function-call-leaderboard/bfcl/model_handler/local_inference/minicpm_fc.py
@@ -3,7 +3,7 @@ import json
 from typing import Dict, List
 
 import datamodel_code_generator
-from bfcl.model_handler.constant import GORILLA_TO_OPENAPI
+from bfcl.model_handler.function_calling_constants import GORILLA_TO_OPENAPI
 from bfcl.model_handler.local_inference.base_oss_handler import OSSHandler
 from bfcl.model_handler.utils import (
     convert_to_tool,
