+++ berkeley-function-call-leaderboard/utils/check_illegal_python_param_name.py
@@ -1,7 +1,7 @@
 from keyword import kwlist
 
 from bfcl._llm_response_generation import parse_test_category_argument
-from bfcl.constant import POSSIBLE_ANSWER_PATH, PROMPT_PATH
+from bfcl.constants import POSSIBLE_ANSWER_PATH, PROMPT_PATH
 from bfcl.utils import (
     is_executable,
     is_java,
